Synthea service no longer runs twice
the startup.sh file ran the synthea service, and then uploader service. but the uploader depends on synthea service service_completed_successfully, per the docker compose. this caused the synthea service to run again when the uploader service stopped. just needed to remove the explicit synthea stuff in startup.sh